:recycle: add section component
